cleanup some more, delete old components
